+++ src/Special Structs/SpinArray.h
@@ -29,7 +29,7 @@ class SpinArray{
         void resetCount();
 
         bool put(short depth, short i, short j);
-        void put(int* frame, int width, int height);
+        void put(short* frame, int width, int height);
         void get(Point *array, int size);
         long getDistance();
         Point* get();
